+++ lib/binance/spot.rb
@@ -6,7 +6,6 @@ require 'binance/utils/validation'
 require 'binance/utils/url'
 require 'binance/error'
 require 'binance/spot/blvt'
-require 'binance/spot/bswap'
 require 'binance/spot/c2c'
 require 'binance/spot/convert'
 require 'binance/spot/fiat'
@@ -15,8 +14,7 @@ require 'binance/spot/loan'
 require 'binance/spot/margin'
 require 'binance/spot/market'
 require 'binance/spot/mining'
-require 'binance/spot/savings'
-require 'binance/spot/staking'
+require 'binance/spot/simple_earn'
 require 'binance/spot/stream'
 require 'binance/spot/subaccount'
 require 'binance/spot/trade'
@@ -26,7 +24,6 @@ require 'binance/spot/websocket'
 module Binance
   # Spot class includes the following modules:
   # - Blvt
-  # - Bswap
   # - C2C
   # - Convert
   # - Fiat
@@ -35,7 +32,7 @@ module Binance
   # - Margin
   # - Market
   # - Mining
-  # - Saving
+  # - SimpleEarn
   # - Stream
   # - Subaccount
   # - Trade
@@ -43,7 +40,6 @@ module Binance
   # @see https://binance-docs.github.io/apidocs/spot/en/
   class Spot
     include Binance::Spot::Blvt
-    include Binance::Spot::Bswap
     include Binance::Spot::C2C
     include Binance::Spot::Convert
     include Binance::Spot::Fiat
@@ -52,8 +48,7 @@ module Binance
     include Binance::Spot::Margin
     include Binance::Spot::Market
     include Binance::Spot::Mining
-    include Binance::Spot::Savings
-    include Binance::Spot::Staking
+    include Binance::Spot::SimpleEarn
     include Binance::Spot::Stream
     include Binance::Spot::Subaccount
     include Binance::Spot::Trade